Various changes to Models
Added a ToString() Method to Model and an implementation to all its inheritants.  This is intended for searching and parsing.
Added a Model type for the other Models to inherit from. This was necessary for creating a ModelList Container, which main purpose so far is to implement a search capability.